fixed issues with random inputs
