+++ .gitignore
@@ -1,0 +1,2 @@
+# ignored the virtual environment folder.
+/shrt/
