+++ requirments.txt
@@ -1,0 +1,6 @@
+certifi==2023.11.17
+charset-normalizer==3.3.2
+idna==3.6
+pyshorteners==1.0.1
+requests==2.31.0
+urllib3==2.1.0
